chore: more Rust 1.95 clippy fixes (unnecessary_cast, unnecessary_sort_by)
- crates/n42-jmt/src/proof.rs:173 — drop redundant `as u8` cast on `key.0[0] >> 4`
  (key.0 is [u8; 32], so the shift result is already u8)
- crates/n42-node/src/mobile_reward.rs:107 — `sort_by(|a, b| a.0.cmp(&b.0))`
  -> `sort_by_key(|a| a.0)`

+++ crates/n42-jmt/src/proof.rs
@@ -170,7 +170,7 @@ mod tests {
         let key = account_key(&addr);
 
         let proof = build_proof(&jmt, key).unwrap().expect("proof should exist");
-        assert_eq!(proof.shard_index, (key.0[0] >> 4) as u8);
+        assert_eq!(proof.shard_index, key.0[0] >> 4);
         assert!(proof.value.is_some());
         assert_eq!(proof.shard_roots.len(), 16);
